mon_projet/requirements.txt
Django>=4.2
djangorestframework
drf-yasg
django-cors-headers
python-docx
reportlab